requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
bardapi==0.1.34
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
colorama==0.4.6
deep-translator==1.11.4
docopt==0.6.2
ffmpeg==1.4
ffmpeg-normalize==0.5.1
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
googletrans==3.0.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.9
lz4==4.3.2
magic-filter==1.0.11
multidict==6.0.4
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.0
typing_extensions==4.7.1
urllib3==2.0.5
yarl==1.9.2
